requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
httpx
protobuf
openai